requirements.txt
streamlit==1.38.0
pydantic==2.8.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-robotexclusionrulesparser==1.7.1
reportlab==4.2.2
pytest==8.3.2